Code du Projet
The objective is to fill a 9x9 grid with numbers so that each column, each row, and each of the nine 3x3 blocks that make up the grid contain all the numbers from 1 to 9. The puzzle creator provides a partially completed grid, which, for a well-posed puzzle has a unique solution. The same integer cannot appear twice in the same row, column or in any of the nine blocks.


You will create this game in command line.
At the beginning, the game asks the user for the level.
Generate and display a feasible and suitable Sudoku grid. You can simply permute rows/columns on a correct existing grid, or transpose; the most ambitious will perform backtracking 
Requests a number and a position to insert the number in the Sudoku grid.
Display the Sudoku grid and ask again.
The game ends when the user has found all the correct numbers in the Sudoku grid.